Big reorganise the core library into `core`
