+++ src/sqlfluff/core/default_config.cfg
