+++ src/sqlfluff/core/dialects/base.py
