+++ src/sqlfluff/core/dialects/dialect_postgres.py
