+++ src/sqlfluff/core/dialects/dialect_teradata.py
