+++ src/sqlfluff/core/errors.py
