+++ src/sqlfluff/core/parser/lexer.py
