+++ src/sqlfluff/core/parser/match_result.py
