+++ src/sqlfluff/core/parser/segment_generator.py
