+++ src/sqlfluff/core/parser/segments_base.py
