+++ src/sqlfluff/core/parser/segments_common.py
