+++ src/sqlfluff/core/rules/base.py
