+++ src/sqlfluff/core/rules/config_info.py
