+++ test/core/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the core library."""
