+++ test/core/linter_test.py
@@ -2,8 +2,8 @@
 
 import pytest
 
-from sqlfluff.linter import Linter, LintingResult
-from sqlfluff.config import FluffConfig
+from sqlfluff.core import FluffConfig, Linter
+from sqlfluff.core.linter import LintingResult
 
 
 def normalise_paths(paths):
