+++ test/core/rules/__init__.py
