+++ test/core/rules/std_roundtrip_test.py
